requirements.txt
pyobjc-framework-Vision
pyobjc-framework-Quartz
pynput
